Removed object-level check for LogEntryEvent__e.isCreateable() (#160)
Previously, if `SObjectType.LogEntryEvent__e.isCreateable()` returned false for a user, no log entries would be saved - I originally put this in place to try to follow best practices for object-level security. However, in many orgs, it's a big effort to get permission sets assigned to users (in this case, the Log Creator permission set). Removing this check will save some headaches in orgs with a lot of existing users and/or more rigid processes for assigning permission sets.