Added new scratch org def file that will only be used for development (not in the pipeline or when creating package versions)
It includes the feature 'EinsteinGPTPlatform', which I've been trying out when writing Apex. It also gives an easy way to enable other features/settings that could be useful for dev, but not needed by any of Nebula Logger's functionality.

Also renamed the existing scratch def files to have a 'build-' prefix

Some org settings that were previously configured via scratch def files are now stored as metadata in a new 'dev' directory (which is now the default directory in sfdx-project.json)